feat: add support for Kerberos
These changes add support for Kerberos authentication; the change to
nodeserver.go is only there to prevent a warning.

To make this work, you need a version of cifs-utils installed on your
nodes (yes, the *nodes*) that contains a fix for CVE-2021-20208, you
need to create a krb5.conf with

  [libdefaults]
  default_ccache_name = /var/krb5/ccache/{%uid}

and put it into a Kubernetes ConfigMap called krb5-conf, and you need
to create a Kubernetes Secret called krb5-keytab with two keys,
"principal", which should name the principal to authenticate with, and
"krb5.keytab", which should be a Kerberos keytab file containing the
key for that principal.